refactor: delete dead shims, split gen-ssg-routes, unify frontmatter parser
Three architecture-cleanup candidates from the recent review:

1. Delete pure-forwarder shims. src/composables/useCoverage.ts forwarded
   fetchCoverage to loadCoverage for 7 callers; the 7 callers now import
   loadCoverage directly from src/lib/unicode/coverage.ts. The four
   clear*Cache() wrappers in fonts/loader.ts, formulas/loader.ts,
   markdown/loader.ts, families-loader.ts were dead code (zero production
   callers); only the test-smoke assertions remain, removed. The two
   wrappers with real test callers (clearMarkdownCache,
   clearDetailsCache) stay — they're the test-accessible seam for the
   singleton caches.

2. Split the gen-ssg-routes.mjs kitchen sink into four purpose-specific
   scripts, matching the pattern already established by
   gen-font-families.mjs and gen-licenses-index.mjs:
     - gen-blog-index.mjs    writes content/blog/index.json
     - gen-guide-index.mjs   writes content/guide/index.json
     - gen-routes.mjs        writes ssg-routes.json
     - gen-sitemap.mjs       walks dist/ AFTER vite-ssg build to write
                              sitemap.xml (more reliable than route-based)
   gen-ssg-routes.mjs remains as a back-compat shim that calls the three
   pre-build scripts in the original order.

3. Unify the frontmatter parser. Build scripts can't import .ts files
   directly, so gen-ssg-routes.mjs had a copy-pasted parseFrontmatterGetters.
   The parser now lives in src/lib/markdown/frontmatter.mjs as the source
   of truth; the .ts file re-exports from .mjs for the runtime. Both
   paths use the same parser.

The build script in package.json now chains: gen-font-families, gen-licenses,
gen-blog-index, gen-guide-index, gen-routes, vite-ssg build, gen-sitemap.
