+++ package.json
@@ -4,12 +4,15 @@
     "dev:astro": "astro dev",
     "build:astro": "astro build",
     "fetch-data": "bash scripts/fetch-data.sh",
-    "gen-routes": "node scripts/gen-ssg-routes.mjs",
+    "gen-routes": "node scripts/gen-blog-index.mjs && node scripts/gen-guide-index.mjs && node scripts/gen-routes.mjs",
+    "gen-sitemap": "node scripts/gen-sitemap.mjs",
     "gen-unicode": "node scripts/gen-unicode-data.mjs",
     "gen-font-families": "node scripts/gen-font-families.mjs",
     "gen-licenses": "node scripts/gen-licenses-index.mjs",
-    "build": "bash scripts/fetch-data.sh && node scripts/gen-font-families.mjs && node scripts/gen-licenses-index.mjs && node scripts/gen-ssg-routes.mjs && vite-ssg build",
-    "build:no-fetch": "node scripts/gen-font-families.mjs && node scripts/gen-licenses-index.mjs && node scripts/gen-ssg-routes.mjs && vite-ssg build",
+    "gen-blog-index": "node scripts/gen-blog-index.mjs",
+    "gen-guide-index": "node scripts/gen-guide-index.mjs",
+    "build": "bash scripts/fetch-data.sh && node scripts/gen-font-families.mjs && node scripts/gen-licenses-index.mjs && node scripts/gen-blog-index.mjs && node scripts/gen-guide-index.mjs && node scripts/gen-routes.mjs && vite-ssg build && node scripts/gen-sitemap.mjs",
+    "build:no-fetch": "node scripts/gen-font-families.mjs && node scripts/gen-licenses-index.mjs && node scripts/gen-blog-index.mjs && node scripts/gen-guide-index.mjs && node scripts/gen-routes.mjs && vite-ssg build && node scripts/gen-sitemap.mjs",
     "preview": "vite preview",
     "format": "prettier -w .",
     "test": "node --test tests/*.test.mjs src/lib/markdown/__tests__/*.test.ts src/lib/unicode/__tests__/*.test.ts src/lib/fonts/__tests__/*.test.ts src/lib/formulas/__tests__/*.test.ts src/lib/__tests__/*.test.ts scripts/__tests__/*.test.ts scripts/__tests__/*.test.mjs",
