+++ scripts/gen-routes.mjs
@@ -1,0 +1,89 @@
+#!/usr/bin/env node
+// gen-routes — reads formulas-data.json + font-families.json +
+// unicode-blocks.json + the blog/guide/license indexes, and writes
+// public/ssg-routes.json with every route the SSG build should render.
+//
+// Companion to gen-sitemap.mjs (which post-processes dist/ into a sitemap),
+// gen-blog-index.mjs (which generates one of this script's inputs), and
+// gen-guide-index.mjs (likewise).
+
+import { readFileSync, readdirSync, writeFileSync } from 'node:fs'
+import { resolve, dirname } from 'node:path'
+import { fileURLToPath } from 'node:url'
+
+const root = resolve(dirname(fileURLToPath(import.meta.url)), '..')
+const pub = resolve(root, 'public')
+
+const STATIC_ROUTES = ['/', '/about', '/blog', '/compare', '/families', '/formulas', '/unicode', '/licenses', '/guide']
+
+const blockToSlug = (name) =>
+  name.toLowerCase().replace(/[^a-z0-9]+/g, '-').replace(/^-|-$/g, '')
+
+function readJson(path, fallback) {
+  try {
+    return JSON.parse(readFileSync(path, 'utf8'))
+  } catch {
+    return fallback
+  }
+}
+
+const formulas = readJson(resolve(pub, 'formulas-data.json'), [])
+const familiesIndex = readJson(resolve(pub, 'font-families.json'), { families: [] })
+const blocksFile = readJson(resolve(pub, 'unicode-blocks.json'), {})
+const blockSlugs = Object.values(blocksFile).map((b) => blockToSlug(b.name)).filter(Boolean)
+const blogIndex = readJson(resolve(pub, 'content/blog/index.json'), [])
+const guideIndex = readJson(resolve(pub, 'content/guide/index.json'), [])
+
+const routes = new Set(STATIC_ROUTES)
+
+for (const f of formulas) {
+  if (!f.slug) continue
+  routes.add(`/formulas/${f.slug}`)
+}
+
+for (const fam of familiesIndex.families || []) {
+  if (!fam.slug) continue
+  routes.add(`/families/${fam.slug}`)
+  routes.add(`/families/${fam.slug}/unicode`)
+}
+
+const emittedFileSlugs = new Set()
+for (const fam of familiesIndex.families || []) {
+  for (const file of fam.files || []) {
+    if (!file.slug || emittedFileSlugs.has(file.slug)) continue
+    emittedFileSlugs.add(file.slug)
+    routes.add(`/fonts/${file.slug}`)
+    routes.add(`/fonts/${file.slug}/unicode`)
+  }
+}
+console.log(`fonts: emitted ${emittedFileSlugs.size} unique file slugs`)
+
+const PLANE_KEYS = ['bmp', 'smp', 'sip', 'tip', 'ssp', 'pua-a', 'pua-b']
+for (const k of PLANE_KEYS) routes.add(`/unicode/plane/${k}`)
+
+const PROPERTY_INDEXES = ['scripts', 'category', 'combining', 'bidiclass']
+for (const p of PROPERTY_INDEXES) routes.add(`/unicode/${p}`)
+
+let licenseSlugs = []
+try {
+  const ls = readdirSync(resolve(pub, 'content/licenses'), { withFileTypes: true })
+  for (const ent of ls) {
+    if (ent.isDirectory()) {
+      licenseSlugs.push(ent.name)
+    } else if (ent.name.endsWith('.md') && ent.name !== 'index.md') {
+      const slug = ent.name.replace(/\.md$/, '')
+      licenseSlugs.push(slug)
+    }
+  }
+} catch {}
+// licenses/index.json is generated by gen-licenses-index.mjs from YAML data.
+// This script only enumerates routes from the directory.
+for (const s of licenseSlugs) routes.add(`/licenses/${s}`)
+for (const s of blockSlugs) routes.add(`/unicode/block/${s}`)
+for (const b of blogIndex) routes.add(`/blog/${b.slug}`)
+for (const g of guideIndex) routes.add(`/guide/${g.slug}`)
+
+const out = resolve(pub, 'ssg-routes.json')
+const routesArr = [...routes].sort()
+writeFileSync(out, JSON.stringify(routesArr, null, 2))
+console.log(`ssg-routes: wrote ${routesArr.length} routes to ${out}`)
